inlining: do not duplicate subparser states in specific case
The goal of this change is to save up the resources needed for parser
implementation.

Before this commit, for each subparser invocation all subparser states
are inlined which may mean that some states are duplicated
unnecessarily.

In a specific case when there is a same sequence of parser statements
and a same transition statement after multiple invocations of the same
subparser instance, it is possible to inline the states only once and
use them for all such invocations.

In this commit, GeneralInliner is changed to inline the subparser
states only for the first such invocation and for all the other
invocations there are used the same states which were already inlined
for that first subparser invocation.

Example testing P4 program is added too.